requirements.txt
# Core dependencies
openai-whisper==20250625
pyannote.audio==3.3.2
python-dotenv==1.1.1
torch>=2.8.0
torchaudio>=2.8.0

# GUI dependencies (tkinter is built into Python)

# Translation dependencies (optional)
googletrans==4.0.2
nest-asyncio

# Required by whisper and pyannote
numpy
tqdm
requests
PyYAML
huggingface-hub
soundfile

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
